Transactions bug fix

+++ app/Http/Controllers/Api/V1/ProfilePhotoController.php
@@ -20,16 +20,6 @@ class ProfilePhotoController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
@@ -55,29 +45,6 @@ class ProfilePhotoController extends Controller
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Image  $image
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Image $image)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Image  $image
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Image $image)
-    {
-        //
-    }
-
-    /**
      * Remove the specified resource from storage.
      *
      * @param  \App\Models\Image  $image
@@ -85,6 +52,7 @@ class ProfilePhotoController extends Controller
      */
     public function destroy(Image $profilePhoto)
     {
+        $this->authorize('delete', $profilePhoto);
         $profilePhoto->delete();
         return response()->noContent();
     }
